make event modal into a component
